Enhance notification and properties services with web push notification support and cache management improvements
